requirements.txt
streamlit
openai
sentence-transformers
google-generativeai
requests
beautifulsoup4
